token: pass the project id to get the catalog
Keystone returns the service catalog only with the scoped token. If the
user has no default project and we don't pass a project_id, we end up
in the case where the token is unscoped.

By passing the project id, we ensure the token is scoped.

Note, the behaviour is not recent, it was already the case OpenStack
Havana. From the changelog:

https://docs.openstack.org/api-ref/identity/v3/#service-catalog-and-endpoints

    A token without an explicit scope of authorization is issued if the user
    does not specify a project and does not have authorization on the project
    specified by their default project attribute

See: https://docs.openstack.org/api-ref/identity/v3/#service-catalog-and-endpoints